+++ src/GCodes/GCodes.h
@@ -241,7 +241,7 @@ public:
 	void EmergencyStop();												// Cancel everything
 	bool GetLastPrintingHeight(float& height) const;					// Get the height in user coordinates of the last printing move
 	bool IsConfigFile(const char* filename);							// Check whether file has .g extension
-	size_t PrintingVolumeNumber(const char *filename);					// Which volume is printing
+	int PrintingVolumeNumber(const char *filename);					// Which volume is printing
 	void CopyFilesFromDir(const StringRef& reply, const char* sourceDir, const char* destinationDir); // Copy files from one directory to another
 	void SavePrintInfoToCSV(const char *filename, StopPrintReason reason, uint32_t hours, uint32_t minutes); // Save print info in csv file
